+++ duckscript_sdk/src/sdk/std/string/mod.rs
@@ -5,6 +5,7 @@ mod indexof;
 mod is_empty;
 mod last_indexof;
 mod starts_with;
+mod substring;
 mod trim;
 mod trim_end;
 mod trim_start;
@@ -25,6 +26,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(is_empty::create(&package))?;
     commands.set(last_indexof::create(&package))?;
     commands.set(starts_with::create(&package))?;
+    commands.set(substring::create(&package))?;
     commands.set(trim::create(&package))?;
     commands.set(trim_start::create(&package))?;
     commands.set(trim_end::create(&package))?;